Set number of rocket workers to 1 (#4)
Currently, the application runs fine locally, but while running in Heroku the application crashes due to too many connections to postgres (max 20 for free plan). It seems like the number of connections in the connection pool defaults to same as the number of rocket workers (threads, concurrent connections, or something like that). So, let's try to limit the number of workers